+++ gradle.properties
@@ -18,4 +18,4 @@ debug_load_all_mods = true
 
 # Coremod Arguments
 include_mod = true
-coremod_plugin_class_name = com.cleanroommc.groovyscript.GroovyScriptCore
+coremod_plugin_class_name = com.cleanroommc.groovyscript.core.GroovyScriptCore
