+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/enderio/AlloySmelter.java
@@ -5,10 +5,10 @@ import com.cleanroommc.groovyscript.compat.mods.ModSupport;
 import com.cleanroommc.groovyscript.compat.mods.enderio.recipe.EnderIORecipeBuilder;
 import com.cleanroommc.groovyscript.compat.mods.enderio.recipe.RecipeInput;
 import com.cleanroommc.groovyscript.helper.ArrayUtils;
-import com.cleanroommc.groovyscript.mixin.enderio.AlloyRecipeManagerAccessor;
-import com.cleanroommc.groovyscript.mixin.enderio.ItemRecipeLeafNodeAccessor;
-import com.cleanroommc.groovyscript.mixin.enderio.ItemRecipeNodeAccessor;
-import com.cleanroommc.groovyscript.mixin.enderio.TriItemLookupAccessor;
+import com.cleanroommc.groovyscript.core.mixin.enderio.AlloyRecipeManagerAccessor;
+import com.cleanroommc.groovyscript.core.mixin.enderio.ItemRecipeLeafNodeAccessor;
+import com.cleanroommc.groovyscript.core.mixin.enderio.ItemRecipeNodeAccessor;
+import com.cleanroommc.groovyscript.core.mixin.enderio.TriItemLookupAccessor;
 import com.cleanroommc.groovyscript.registry.VirtualizedRegistry;
 import com.cleanroommc.groovyscript.sandbox.GroovyLog;
 import com.enderio.core.common.util.NNList;
