+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/thermalexpansion/Brewer.java
@@ -8,7 +8,7 @@ import com.cleanroommc.groovyscript.api.IIngredient;
 import com.cleanroommc.groovyscript.compat.EnergyRecipeBuilder;
 import com.cleanroommc.groovyscript.compat.mods.ModSupport;
 import com.cleanroommc.groovyscript.helper.IngredientHelper;
-import com.cleanroommc.groovyscript.mixin.thermalexpansion.BrewerManagerAccessor;
+import com.cleanroommc.groovyscript.core.mixin.thermalexpansion.BrewerManagerAccessor;
 import com.cleanroommc.groovyscript.registry.VirtualizedRegistry;
 import com.cleanroommc.groovyscript.sandbox.GroovyLog;
 import net.minecraft.item.ItemStack;
