+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/thermalexpansion/Crucible.java
@@ -7,7 +7,7 @@ import com.cleanroommc.groovyscript.api.IIngredient;
 import com.cleanroommc.groovyscript.compat.EnergyRecipeBuilder;
 import com.cleanroommc.groovyscript.compat.mods.ModSupport;
 import com.cleanroommc.groovyscript.helper.IngredientHelper;
-import com.cleanroommc.groovyscript.mixin.thermalexpansion.CrucibleManagerAccessor;
+import com.cleanroommc.groovyscript.core.mixin.thermalexpansion.CrucibleManagerAccessor;
 import com.cleanroommc.groovyscript.registry.VirtualizedRegistry;
 import com.cleanroommc.groovyscript.sandbox.GroovyLog;
 import net.minecraft.item.ItemStack;
