+++ src/main/java/com/cleanroommc/groovyscript/compat/mods/thermalexpansion/Pulverizer.java
@@ -9,7 +9,7 @@ import com.cleanroommc.groovyscript.compat.EnergyRecipeBuilder;
 import com.cleanroommc.groovyscript.compat.mods.ModSupport;
 import com.cleanroommc.groovyscript.helper.IngredientHelper;
 import com.cleanroommc.groovyscript.helper.SimpleObjectStream;
-import com.cleanroommc.groovyscript.mixin.thermalexpansion.PulverizerManagerAccessor;
+import com.cleanroommc.groovyscript.core.mixin.thermalexpansion.PulverizerManagerAccessor;
 import com.cleanroommc.groovyscript.registry.VirtualizedRegistry;
 import com.cleanroommc.groovyscript.sandbox.GroovyLog;
 import net.minecraft.item.ItemStack;
