+++ src/main/java/com/cleanroommc/groovyscript/core/mixin/enderio/AlloyRecipeManagerAccessor.java
@@ -1,4 +1,4 @@
-package com.cleanroommc.groovyscript.mixin.enderio;
+package com.cleanroommc.groovyscript.core.mixin.enderio;
 
 import crazypants.enderio.base.recipe.IManyToOneRecipe;
 import crazypants.enderio.base.recipe.alloysmelter.AlloyRecipeManager;
