+++ src/main/java/com/cleanroommc/groovyscript/core/mixin/thermalexpansion/CrucibleManagerAccessor.java
@@ -1,4 +1,4 @@
-package com.cleanroommc.groovyscript.mixin.thermalexpansion;
+package com.cleanroommc.groovyscript.core.mixin.thermalexpansion;
 
 import cofh.core.inventory.ComparableItemStackValidatedNBT;
 import cofh.thermalexpansion.util.managers.machine.CrucibleManager;
