+++ src/main/resources/mixin.groovyscript.enderio.json
@@ -1,5 +1,5 @@
 {
-  "package": "com.cleanroommc.groovyscript.mixin.enderio",
+  "package": "com.cleanroommc.groovyscript.core.mixin.enderio",
   "refmap": "mixins.bogosorter.refmap.json",
   "target": "@env(DEFAULT)",
   "minVersion": "0.8",
