+++ src/main/resources/mixin.groovyscript.jei.json
@@ -1,5 +1,5 @@
 {
-  "package": "com.cleanroommc.groovyscript.mixin.jei",
+  "package": "com.cleanroommc.groovyscript.core.mixin.jei",
   "refmap": "mixins.bogosorter.refmap.json",
   "target": "@env(DEFAULT)",
   "minVersion": "0.8",
