+++ src/main/resources/mixin.groovyscript.json
@@ -1,5 +1,5 @@
 {
-  "package": "com.cleanroommc.groovyscript.mixin",
+  "package": "com.cleanroommc.groovyscript.core.mixin",
   "refmap": "mixins.bogosorter.refmap.json",
   "target": "@env(DEFAULT)",
   "minVersion": "0.8",
