+++ src/main/resources/mixin.groovyscript.mekanism.json
@@ -1,5 +1,5 @@
 {
-  "package": "com.cleanroommc.groovyscript.mixin.mekanism",
+  "package": "com.cleanroommc.groovyscript.core.mixin.mekanism",
   "refmap": "mixins.bogosorter.refmap.json",
   "target": "@env(DEFAULT)",
   "minVersion": "0.8",
